Moved Synth Project to its own repository
Started creating functions that generate wave tables and also some functions that smooth out the sound via interpolation and fade in/out. Can switch between sin wave and a sawtooth wave. Want to look for more formulas for different wave tables. Was able to generate wav files and tune to different notes in the scale of C Major.